+++ packages/zowe-explorer-api/src/profiles/ZoweExplorerZosmfApi.ts
@@ -368,8 +368,8 @@ export class ZosmfJesApi extends ZosmfApiCommon implements ZoweExplorerApi.IJes 
         return zowe.DownloadJobs.downloadSpoolContentCommon(this.getSession(), parms);
     }
 
-    public getSpoolContentById(jobname: string, jobid: string, spoolId: number): Promise<string> {
-        return zowe.GetJobs.getSpoolContentById(this.getSession(), jobname, jobid, spoolId);
+    public getSpoolContentById(jobname: string, jobid: string, spoolId: number, encoding?: string): Promise<string> {
+        return zowe.GetJobs.getSpoolContentById(this.getSession(), jobname, jobid, spoolId, encoding);
     }
 
     public getJclForJob(job: zowe.IJob): Promise<string> {